pyproject.toml
[project]
name = "scraping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiofiles>=25.1.0",
    "aiosqlite>=0.22.1",
    "ftfy>=6.3.1",
    "loguru>=0.7.3",
    "markdownify>=1.2.2",
    "markitdown[pdf]>=0.1.4",
    "nh3>=0.3.2",
    "patchright>=1.57.2",
    "polars>=1.37.1",
    "pydantic>=2.12.5",
    "pytest-asyncio>=1.3.0",
    "pyyaml>=6.0.3",
    "questionary>=2.1.1",
    "rich>=14.2.0",
    "scrapling>=0.3.14",
    "selectolax>=0.4.6",
    "trafilatura>=2.0.0",
]

[project.scripts]
uif-scraper = "uif_scraper.cli:main"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.13",
    "types-aiofiles>=25.1.0.20251011",
    "types-pyyaml>=6.0.12.20250915",
]
